Create handle the last state.go
To handle the last state stored as vector commitments in the Ethereum light client, you would need to modify the code to include the verification of vector commitments in addition to the Merkle proof validation
In this updated code, a new function called verifyVectorCommitmentProof is added to perform the verification of the vector commitment proof. This function uses an RPC call (e.g., transition_getProof) to retrieve the transition proof associated with the provided proof root and then verifies the proof using the utils.VerifyProof function.

The verifyState function is updated to include the verification of the vector commitment proof by calling verifyVectorCommitmentProof with the appropriate arguments.

Note that the verifyVectorCommitmentProof function is just an example implementation. You will need to replace it with the actual logic and library or RPC call that verifies the vector commitment proof for the Ethereum blockchain you are working with.