Update and rename light client.go to stateless-light client.go
In the updated code, a new struct HeaderWithAggregatedSubvector is defined, which extends the existing types. Header struct and includes an additional field AggregatedSubvector of type common.Hash. This new field represents the aggregated subvector commitment.

The NewHeaderWithAggregatedSubvector function is also added to create instances of HeaderWithAggregatedSubvector. It takes all the existing fields of types.Header along with the new aggregatedSubvector parameter to initialize the struct.

By using HeaderWithAggregatedSubvector instead of types. Header, we can include the aggregated subvector commitment in the header of the stateless light client.